{"schema":"decodex/commit/1","summary":"Extract frozen annotation style models","authority":"manual"}
